Generator/Client/requirements.txt
google-cloud-pubsub
Faker==8.8.1
google-auth